@@ -1,4 +1,4 @@
-class Place < ActiveRecord::Base
+class Place < ApplicationRecord
   belongs_to :tournament
   belongs_to :division
   belongs_to :team
