@@ -1,4 +1,4 @@
-class PoolResult < ActiveRecord::Base
+class PoolResult < ApplicationRecord
   belongs_to :tournament
   belongs_to :division
   belongs_to :team
