@@ -1,4 +1,4 @@
-class ScoreReport < ActiveRecord::Base
+class ScoreReport < ApplicationRecord
   belongs_to :tournament
   belongs_to :game
   belongs_to :team
