@@ -1,4 +1,4 @@
-class User < ActiveRecord::Base
+class User < ApplicationRecord
   has_many :user_authentications, dependent: :destroy
   has_many :tournament_users, dependent: :destroy
   has_many :tournaments, through: :tournament_users
