@@ -22,7 +22,7 @@ class Admin::SettingsController < AdminController
       :time_cap
     )
 
-    tournament_params.merge!(timezone: Time.zone.name)
+    tournament_params[:timezone] = Time.zone.name
     tournament_params
   end
 end
