@@ -1,5 +1,5 @@
 module Teams
-  class DeleteScoreReportsJob < ActiveJob::Base
+  class DeleteScoreReportsJob < ApplicationJob
     queue_as :default
 
     def perform(tournament_id:, team_id:)
