@@ -30,7 +30,7 @@ class AdminTest < ActionDispatch::IntegrationTest
     assert_equal 200, status
     assert_equal new_user_session_path, path
 
-    post new_user_session_path, user: {email: user.email, password: 'password'}
+    post new_user_session_path, params: { user: {email: user.email, password: 'password'} }
     follow_redirect!
     assert_equal 200, status
     assert_equal "/admin", path
