@@ -11,7 +11,7 @@ class LoginTest < ActionDispatch::IntegrationTest
     assert_equal 200, status
     assert_equal new_user_session_path, path
 
-    post new_user_session_path, user: {email: @user.email, password: 'password'}
+    post new_user_session_path, params: { user: {email: @user.email, password: 'password'} }
     follow_redirect!
     assert_equal 200, status
     assert_equal "/admin", path
@@ -23,7 +23,7 @@ class LoginTest < ActionDispatch::IntegrationTest
     assert_equal 200, status
     assert_equal new_user_session_path, path
 
-    post new_user_session_path, user: {email: @user.email, password: 'password'}
+    post new_user_session_path, params: { user: {email: @user.email, password: 'password'} }
     follow_redirect!
     assert_equal 200, status
     assert_equal "/admin", path
@@ -38,7 +38,7 @@ class LoginTest < ActionDispatch::IntegrationTest
     assert_equal 200, status
     assert_equal new_user_session_path, path
 
-    post new_user_session_path, user: {email: @user.email, password: 'password'}
+    post new_user_session_path, params: { user: {email: @user.email, password: 'password'} }
     assert_equal 200, status
     assert_equal new_user_session_path, path
   end
@@ -49,7 +49,7 @@ class LoginTest < ActionDispatch::IntegrationTest
     assert_equal 200, status
     assert_equal new_user_session_path, path
 
-    post new_user_session_path, user: {email: @user.email, password: 'password'}
+    post new_user_session_path, params: {user: {email: @user.email, password: 'password'} }
     follow_redirect!
     assert_equal 200, status
     assert_equal "/admin/fields", path
