@@ -18,7 +18,7 @@ end
 
 class BrowserTest < ActiveSupport::TestCase
   include Capybara::DSL
-  self.use_transactional_fixtures = false
+  self.use_transactional_tests = false
 
   setup do
     @user = users(:kevin)
